+++ Loan_management/backend/templates/login.html
@@ -1,26 +1,16 @@
 <div id="login_form" class="d-flex align-items-center justify-content-center" style="height: 275px;">
-<form class="px-4 py-3">
+<form class="px-4 py-3" action="/users/login" method="POST">
+  {%csrf_token%}
       <div class="form-group">
-        <input type="text" class="form-control" id="username" placeholder="User name">
+        <input type="text" class="form-control" name="username" id="username" placeholder="User name">
       </div>
       <div class="form-group">
-        <input type="password" class="form-control" id="pwd" placeholder="Password">
+        <input type="password" class="form-control" name="pwd" id="pwd" placeholder="Password">
       </div>
-      <div class="form-row align-items-center">
-        <div class="col-auto my-1">
-          <select class="custom-select mr-sm-2" id="user_type">
-            <option selected>User type</option>
-            <option value="1">User</option>
-            <option value="2">Admin</option>
-            <option value="3">Manager</option>
-            <option value="4">Auditor</option>
-          </select>
-        </div>
-     </div>
       <button type="submit" class="btn btn-primary">Sign in</button>
     </form>
 </div>
-<div class="w-25 mx-auto"><a href="/web_site/signup" class="link-dark" >New around here? Sign up</a></div>
+<div class="w-25 mx-auto"><a href="/users/signup" class="link-dark" >New around here? Sign up</a></div>
 <div class="w-25 mx-auto"><a href="#" class="link-danger" >Forgot password?</a></div>
