+++ Loan_management/backend/templates/roles.html
@@ -11,6 +11,7 @@
                 <tr><td>{{r.id}}</td><td>{{r.role_name}}</td></tr>
             {%endfor%}
         </table>
+        <a href="/role/add"></a>
         <footer>{%include "footer.html"%}</footer>
     </body>
