--- Loan_management/backend/templates/signup.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-  </head>
-  <body>
-    <header>{%include "header.html"%}</header>
-    <div id="login_form" class="d-flex align-items-center justify-content-center" style="height: 275px;">
-    <form class="px-4 py-3">
-          <div class="form-group">
-            <input type="text" class="form-control" id="new_username" placeholder="New user name">
-          </div>
-          <div class="form-group">
-            <input type="password" class="form-control" id="pwd" placeholder="Password">
-          </div>
-          <div class="form-group">
-            <input type="password" class="form-control" id="cnf_pwd" placeholder="Confirm password">
-          </div>
-          <button type="submit" class="btn btn-primary">Save</button>
-        </form>
-    </div>
-<footer>{%include "footer.html"%}</footer>
-  </body>
-</html>
-    
-        
-
