+++ Loan_management/loans/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class LoansConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'loans'
