+++ Loan_management/loans/urls.py
@@ -1,0 +1,4 @@
+from django.urls import path
+from . import views
+
+urlpatterns = []
