+++ Loan_management/payments/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class WebSiteConfig(AppConfig):
+class PaymentsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'web_site'
+    name = 'payments'
